Split running application from Prices similar to Typescript.
